+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added `Awesomite\Chariot\Pattern\PatternInterface` - possibility to conversions url params, e.g. date in format YYYY-mm-dd to \DateTime object
 * Added [behat] tests
+* Force pattern names prefixed by ":"
 
 ### [0.1.0] 2017-07-20